Update ModelUFS for param/config changes in oops develop
