+++ src/soca/Model/ufs/ModelUFS.cc
@@ -11,6 +11,8 @@
 
 #include "oops/util/DateTime.h"
 #include "oops/util/Logger.h"
+#include "oops/util/parameters/Parameters.h"
+#include "oops/util/parameters/RequiredParameter.h"
 
 #include "ModelUFS.interface.h"
 
@@ -21,19 +23,34 @@
 
 namespace soca {
 // -------------------------------------------------------------------------------------------------
+// Options taken by ModelUFS
+class ModelUFSParameters : public oops::ModelParametersBase {
+  OOPS_CONCRETE_PARAMETERS(ModelUFSParameters, ModelParametersBase)
+
+ public:
+  oops::RequiredParameter<oops::Variables> modelVariables{ "model variables", this};
+  oops::RequiredParameter<util::Duration> tstep{ "tstep", this};
+  oops::RequiredParameter<std::string> ufsRunDirectory{ "ufs_run_directory", this};
+};
+// -------------------------------------------------------------------------------------------------
 static oops::interface::ModelMaker<Traits, ModelUFS> makermodel_("UFS");
 // -------------------------------------------------------------------------------------------------
-ModelUFS::ModelUFS(const Geometry & resol, const Parameters_ & params)
-  : keyConfig_(0), tstep_(0), geom_(resol), vars_(params.modelVariables) {
+ModelUFS::ModelUFS(const Geometry & resol, const eckit::Configuration & config)
+  : keyConfig_(0), tstep_(0), geom_(resol), vars_() {
   char tmpdir_[10000];
   oops::Log::trace() << "ModelUFS::ModelUFS starting" << std::endl;
   getcwd(tmpdir_, 10000);
+
+  ModelUFSParameters params;
+  params.deserialize(config);
+  vars_ = params.modelVariables;
   tstep_ = params.tstep;
   strcpy(ufsdir_, params.ufsRunDirectory.value().c_str());
   chdir(ufsdir_);
-  soca_ufs_create_f90(keyConfig_, params.toConfiguration(), geom_.toFortran());
-  oops::Log::trace() << "ModelUFS::ModelUFS done" << std::endl;
+  soca_ufs_create_f90(keyConfig_, config, geom_.toFortran());
+
   chdir(tmpdir_);
+  oops::Log::trace() << "ModelUFS::ModelUFS done" << std::endl;
 }
 // -------------------------------------------------------------------------------------------------
 ModelUFS::~ModelUFS() {
