+++ src/soca/Model/ufs/ModelUFS.h
@@ -10,16 +10,13 @@
 #include <ostream>
 #include <string>
 
-#include "oops/base/ParameterTraitsVariables.h"
+#include "eckit/config/Configuration.h"
+
 #include "oops/base/Variables.h"
 #include "oops/generic/ModelBase.h"
 #include "oops/interface/ModelBase.h"
 #include "oops/util/Duration.h"
 #include "oops/util/ObjectCounter.h"
-#include "oops/util/parameters/OptionalParameter.h"
-#include "oops/util/parameters/Parameter.h"
-#include "oops/util/parameters/Parameters.h"
-#include "oops/util/parameters/RequiredParameter.h"
 #include "oops/util/Printable.h"
 
 #include "soca/Geometry/Geometry.h"
@@ -38,26 +35,13 @@ namespace soca {
 
 // -------------------------------------------------------------------------------------------------
 // -------------------------------------------------------------------------------------------------
-/// Options taken by ModelUFS
-  class ModelUFSParameters : public oops::ModelParametersBase {
-    OOPS_CONCRETE_PARAMETERS(ModelUFSParameters, ModelParametersBase)
-
-   public:
-    oops::RequiredParameter<oops::Variables> modelVariables{ "model variables", this};
-    oops::RequiredParameter<util::Duration> tstep{ "tstep", this};
-    oops::RequiredParameter<std::string> ufsRunDirectory{ "ufs_run_directory", this};
-  };
-
-// -------------------------------------------------------------------------------------------------
-// -------------------------------------------------------------------------------------------------
 
 class ModelUFS: public oops::interface::ModelBase<Traits>,
                 private util::ObjectCounter<ModelUFS> {
  public:
-  typedef ModelUFSParameters Parameters_;
   static const std::string classname() {return "soca::ModelUFS";}
 
-  ModelUFS(const Geometry &, const Parameters_ &);
+  ModelUFS(const Geometry &, const eckit::Configuration &);
   ~ModelUFS();
 
   void initialize(State &) const;
